+++ apps/blog/web/src/app/posts/[...slug]/PostClient.tsx
@@ -10,8 +10,8 @@ import { SsgoiTransition } from '@ssgoi/react';
 
 import type { PostData } from '@/domain/post';
 import GiscusComments from '@/src/components/GiscusComments';
-import { useViewCount } from '@/lib/hooks/useViewCount';
-import { useRecordRecentView } from '@/lib/hooks/useRecentViews';
+import { useViewCount } from '@/src/hooks/useViewCount';
+import { useRecordRecentView } from '@/src/hooks/useRecentViews';
 import { BackToTop } from '@/src/components/mobile/BackToTop';
 import { MobileTOC } from '@/src/components/mobile/MobileTOC';
 import { ShareButton } from '@/src/components/mobile/ShareButton';
